+++ CHANGELOG.md
@@ -74,6 +74,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [20.8.4] - Unreleased
 ### Added
+- Add function to duplicate host and vhost objects [#592](https://github.com/greenbone/gvm-libs/pull/592)
 ### Changed
 * Use better defaults for installation directories [#574](https://github.com/greenbone/gvm-libs/pull/574)
     * SYSCONFDIR is /etc by default now